Allow conversion of `Rectangle` to `Vector` and from `Vector` to scalar `Rectangle`
The `toVector` method can convert a rectangle object to a vector object.
To access members of the `Vector` class from the `Rectangle` class, the
`toVector` method must be called first.

Allows a Rectangle object which is more functional than a Vector
object to be converted to a Vector when vector attributes are required
from the Rectangle
which is scalar