check that Drawer state passed through, and qualified for a `touchmove` event before going through `touchend`
The touchmove doesn't qualify until the touch point is less than/equal
to the width of the Drawer element, on close. So until this condition
evals. `true` a touchmove doesn't qualify, and why should a touchend
qualify too?

This prevents closing the Drawer element on touchend when the touchmove
handler didn't call. For example a touchmove that occurs on the
backdrop and not the Drawer element will close the Drawer element as
long as touchmove hits threshold, although on the wrong target.