Reverted
Tried to make the Rectangle which is Scalar convert to Vector and
block access to Vector properties and methods on default, though it
extends Vector. The Vector props. and methods will be available when
a `toVector` is called. This resulted in getting
"bugs as much as characters in this code". Hence the change was reverted