improved API
A navigation object is constructed by calling `new Nav()` the object is
setup using the `setup` method, passing necessary parameters.
The setup method returns a `NavMountWorker` object. The `mount`
method of this object must be called to mount a Navigation activity
and activate necessary services for the activity to run. The `unmount`
method is used to unmount the activity and deactivate all associated
services, even stubborn services are forced to deactivate.